[Item] CAES-1023: Refactor Input ItemField component
