--- packages/components/FormInput/FormInput.js
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-import { Input } from '../Input';
-
-const FormInput = styled(Input)`
-  ${Input.InputField} {
-    position: relative;
-    padding: 5px 15px;
-    color: ${({ theme }) => theme.color.black};
-  }
-`;
-
-export default FormInput;
