--- packages/components/FormInput/index.js
@@ -1,1 +1,0 @@
-export { default as FormInput } from './FormInput';
