+++ packages/components/Input/PasswordInput.js
@@ -1,6 +1,6 @@
 import React, { PureComponent } from 'react';
 import styled from 'styled-components';
-import Input from './Input';
+import { Input } from './Input';
 import { Icon } from '../Icon';
 import { HoldClickBehaviour } from '../HoldClickBehaviour';
