+++ packages/components/index.js
@@ -19,12 +19,12 @@ export { DottedMenu } from './DottedMenu';
 export { Dropdown } from './Dropdown';
 export { Error } from './Error';
 export { File } from './File';
-export { FormInput } from './FormInput';
 export { Head } from './Head';
 export { Hint } from './Hint';
 export { HoldClickBehaviour } from './HoldClickBehaviour';
 export { Icon } from './Icon';
 export {
+  FormInput,
   LockInput,
   Input,
   PasswordInput,
