+++ packages/secure-app/package.json
@@ -46,7 +46,6 @@
     "jszip": "^3.2.2",
     "localforage": "^1.7.3",
     "lodash.debounce": "^4.0.8",
-    "lodash.groupby": "^4.6.0",
     "lodash.throttle": "^4.1.1",
     "luxon": "^1.17.1",
     "memoize-one": "^5.0.5",
